backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-vite==3.1.0
djangorestframework==3.16.0
sqlparse==0.5.3
